requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
googleapis-common-protos==1.63.0
requests==2.32.3
python-dateutil>=2.8.2

